refactor (post): move form tags to separate component
